requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Async HTTP
httpx==0.28.1
aiohttp==3.11.11

# Database
motor==3.7.0
pymongo==4.10.1

# Task queue
celery[redis]==5.4.0
redis==5.2.1

# Scraping
scrapy==2.12.0
playwright==1.49.1
beautifulsoup4==4.12.3
lxml==5.3.0

# Email validation
dnspython==2.7.0

# Settings
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Utilities
tenacity==9.0.0
